benchmark/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
anyio==4.7.0
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.2.0
cattrs==24.1.2
certifi==2024.8.30
charset-normalizer==3.4.0
comm==0.2.2
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docstring-to-markdown==0.15
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.3
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
jedi-language-server==0.41.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonrpcserver==5.0.9
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
Levenshtein==0.26.1
lsprotocol==2023.0.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.1.0
multilspy==0.0.1
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest_asyncio==1.6.0
numpy==2.1.2
openai==1.57.4
OSlash==0.6.3
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pip==24.2
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.1
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==1.10.5
pygls==1.3.1
Pygments==2.18.0
pytest==8.3.4
pytest-asyncio==0.24.0
python-dateutil==2.9.0
python-dotenv==1.0.1
python-engineio==4.10.1
python-Levenshtein==0.26.1
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.0
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
pyzmq==25.1.2
RapidFuzz==3.11.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.3
safetensors==0.4.5
setuptools==75.1.0
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.2
tinycss2==1.4.0
tokenizers==0.20.3
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.3
tree-sitter==0.23.2
tree-sitter-java==0.23.2
tree-sitter-python==0.23.2
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.1
wheel==0.44.0
wsproto==1.2.0
yarl==1.18.3
zipp==3.20.2
